pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "geo-maps-agents"
version = "0.1.0"
description = "Production-flavored OpenAI Agents SDK reference implementation for multi-stage geospatial trip planning"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "hafnium49", email = ""},
]
keywords = ["openai", "agents", "geospatial", "google-maps", "trip-planning", "h3", "routing"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "openai-agents>=0.4.0,<0.5.0",
    "httpx>=0.27.0,<0.28.0",
    "h3>=3.7.6,<4.0.0",
    "hdbscan>=0.8.33,<0.9.0",
    "ortools>=9.10.0,<10.0.0",
    "pydantic>=2.0.0,<3.0.0",
    "cachetools>=5.3.0,<6.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "uvicorn[standard]>=0.30.0,<0.31.0",
    "numpy>=1.24.0,<2.0.0",
    "pandas>=2.0.0,<3.0.0",
    "pyyaml>=6.0.0,<7.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0,<9.0.0",
    "pytest-cov>=5.0.0,<6.0.0",
    "pytest-asyncio>=0.23.0,<0.24.0",
    "ruff>=0.5.0,<0.6.0",
    "mypy>=1.10.0,<2.0.0",
    "black>=24.0.0,<25.0.0",
]

[project.scripts]
geo-trip = "geotrip_agent:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "geotrip_agent"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.black]
line-length = 120
target-version = ["py310", "py311", "py312"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --cov=src --cov-report=term-missing --cov-report=html"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
